Rename to viirs_edr_active_fires
