+++ satpy/etc/readers/viirs_edr_active_fires.yaml
