+++ satpy/tests/reader_tests/__init__.py
@@ -38,7 +38,7 @@ from satpy.tests.reader_tests import (test_abi_l1b, test_hrit_base,
                                       test_nc_slstr, test_olci_nc,
                                       test_viirs_edr_flood, test_nwcsaf_nc,
                                       test_seviri_l1b_hrit, test_sar_c_safe,
-                                      test_modis_l1b, test_viirs_edr_fires)
+                                      test_modis_l1b, test_viirs_edr_active_fires)
 
 
 if sys.version_info < (2, 7):
@@ -82,6 +82,6 @@ def suite():
     mysuite.addTests(test_seviri_l1b_hrit.suite())
     mysuite.addTests(test_sar_c_safe.suite())
     mysuite.addTests(test_modis_l1b.suite())
-    mysuite.addTests(test_viirs_edr_fires.suite())
+    mysuite.addTests(test_viirs_edr_active_fires.suite())
 
     return mysuite
