+++ satpy/tests/reader_tests/test_viirs_edr_active_fires.py
@@ -119,7 +119,7 @@ class TestVIIRSActiveFiresNetCDF4(unittest.TestCase):
     def setUp(self):
         """Wrap CDF4 file handler with own fake file handler"""
         from satpy.config import config_search_paths
-        from satpy.readers.viirs_edr_fires import VIIRSActiveFiresFileHandler
+        from satpy.readers.viirs_edr_active_fires import VIIRSActiveFiresFileHandler
         self.reader_configs = config_search_paths(os.path.join('readers', self.yaml_file))
         self.p = mock.patch.object(VIIRSActiveFiresFileHandler, '__bases__', (FakeFiresNetCDF4FileHandler,))
         self.fake_handler = self.p.start()
@@ -171,7 +171,7 @@ class TestVIIRSActiveFiresText(unittest.TestCase):
     def setUp(self):
         """Wrap file handler with own fake file handler"""
         from satpy.config import config_search_paths
-        from satpy.readers.viirs_edr_fires import VIIRSActiveFiresTextFileHandler
+        from satpy.readers.viirs_edr_active_fires import VIIRSActiveFiresTextFileHandler
         self.reader_configs = config_search_paths(os.path.join('readers', self.yaml_file))
         self.p = mock.patch.object(VIIRSActiveFiresTextFileHandler, '__bases__', (FakeFiresTextFileHandler,))
         self.fake_handler = self.p.start()
